backend/requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
fastapi==0.95.1
filelock==3.12.0
h11==0.14.0
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
lit==16.0.1
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
Pillow==9.5.0
pydantic==1.10.7
pypika-tortoise==0.1.6
python-multipart==0.0.6
pytz==2023.3
requests==2.28.2
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
torch==2.0.0
torchvision==0.15.1
tortoise-orm==0.19.3
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
